[Chore] #90 카페 리스트 화면, 필터 하단시트뷰 관련 UI, 모델 이름 및 파일명 정리
